refactor: Util 사용 위치 변경
계층 책임 분리
