+++ app/src/main/java/kr/com/duri/groomer/application/facade/ShopAuthFacade.java
@@ -1,6 +1,5 @@
 package kr.com.duri.groomer.application.facade;
 
-import kr.com.duri.common.security.jwt.JwtUtil;
 import kr.com.duri.groomer.application.dto.response.NewShopJwtResponse;
 import kr.com.duri.groomer.application.mapper.ShopJwtMapper;
 import kr.com.duri.groomer.application.service.ShopService;
@@ -15,8 +14,6 @@ public class ShopAuthFacade {
 
     private final ShopService shopService;
 
-    private final JwtUtil jwtUtil;
-
     private final ShopJwtMapper shopJwtMapper;
 
     public NewShopJwtResponse createNewShopJwt(String providerId) {
@@ -25,8 +22,8 @@ public class ShopAuthFacade {
                         .findBySocialId(providerId)
                         .orElseThrow(() -> new IllegalArgumentException("사용자를 찾을 수 없습니다."));
 
-        String token = jwtUtil.createJwt(shop.getId(), providerId, 60 * 60 * 60 * 60L);
+        String token = shopService.createNewShopJwt(shop);
 
-        return shopJwtMapper.toNewJwtResponse("authorization_shop", token, shop.getNewShop());
+        return shopJwtMapper.toNewJwtResponse(shop, token);
     }
 }
