+++ app/src/main/java/kr/com/duri/groomer/application/mapper/ShopJwtMapper.java
@@ -2,12 +2,13 @@ package kr.com.duri.groomer.application.mapper;
 
 import kr.com.duri.groomer.application.dto.response.NewShopJwtResponse;
 
+import kr.com.duri.groomer.domain.entity.Shop;
 import org.springframework.stereotype.Component;
 
 @Component
 public class ShopJwtMapper {
 
-    public NewShopJwtResponse toNewJwtResponse(String client, String token, Boolean newUser) {
-        return NewShopJwtResponse.builder().token(token).client(client).newUser(newUser).build();
+    public NewShopJwtResponse toNewJwtResponse(Shop shop, String token) {
+        return NewShopJwtResponse.builder().token(token).client("authorization_shop").newUser(shop.getNewShop()).build();
     }
 }
