+++ app/src/main/java/kr/com/duri/user/application/facade/UserAuthFacade.java
@@ -1,6 +1,5 @@
 package kr.com.duri.user.application.facade;
 
-import kr.com.duri.common.security.jwt.JwtUtil;
 import kr.com.duri.user.application.dto.response.NewUserJwtResponse;
 import kr.com.duri.user.application.mapper.UserJwtMapper;
 import kr.com.duri.user.application.service.SiteUserService;
@@ -15,8 +14,6 @@ public class UserAuthFacade {
 
     private final SiteUserService siteUserService;
 
-    private final JwtUtil jwtUtil;
-
     private final UserJwtMapper UserJwtMapper;
 
     public NewUserJwtResponse createNewUserJwt(String providerId) {
@@ -25,9 +22,8 @@ public class UserAuthFacade {
                         .findBySocialId(providerId)
                         .orElseThrow(() -> new IllegalArgumentException("사용자를 찾을 수 없습니다."));
 
-        String token = jwtUtil.createJwt(siteUser.getId(), providerId, 60 * 60 * 60 * 60L);
+        String token = siteUserService.createNewUserJwt(siteUser);
 
-        return UserJwtMapper.toNewUserJwtResponse(
-                siteUser.getName(), "authorization_user", token, siteUser.getNewUser());
+        return UserJwtMapper.toNewUserJwtResponse(siteUser, token);
     }
 }
