+++ app/src/main/java/kr/com/duri/user/application/service/SiteUserService.java
@@ -15,4 +15,6 @@ public interface SiteUserService {
             String gender,
             String birth,
             String birthYear);
+
+    String createNewUserJwt(SiteUser siteUser);
 }
